[IMP] im_livechat, *: split public livechat templates per widget
*: website_livechat

The goal is to split the public_livechat.xml file into multiple parts,
and move each template into its own file located in
`im_livechat/static/src/legacy/widgets/`

Task-2825235

closes odoo/odoo#92671
