+++ addons/im_livechat/static/src/legacy/widgets/document_viewer/document_viewer.js
