+++ addons/im_livechat/static/src/legacy/widgets/feedback/feedback.js
